+++ mapmatching/match/viterbi.py
@@ -70,8 +70,7 @@ def find_matched_sequence(cands, gt, net, dir_trans=True, mode='*', trim_factor=
     gt_beam = pd.concat(gt_beam)
     ratio = gt_beam.shape[0] / gt.shape[0]
     _log = f"Route planning time cost: {np.sum(times):.3f} s, trim ratio: {(1 - ratio)*100:.1f} %"
-    print(_log)
-    # getattr(logger, level)(_log)
+    getattr(logger, level)(_log)
     
     return end_prob, rList, gt_beam
 
@@ -180,8 +179,6 @@ def prepare_viterbi_input(cands, gt):
 
     # BUG cands 坐标不连续的问题, 莫非是中断
     trans_prob = [gt.loc[i]['f'].to_dict() for i in observations[:-1] ]
-    
-    # tmp = gt[['d_euc', 'first_step_len', 'last_step_len', 'cost', 'path', 'geometry', 'v', 'first_step', 'last_step', 'move_dir', 'dir_prob']]
     
     return states, observations, start_prob, trans_prob, emit_prob