+++ mapmatching/osmnet/downloader.py
@@ -50,8 +50,8 @@ if __name__ == "__main__":
     sys.path.append('..')
     from setting import GBA_BBOX, SZ_BBOX
 
+    download_osm_xml('/home/pcl/minio/geo_data/Shenzhen.osm.xml', SZ_BBOX)
     download_osm_xml('../../cache/Futian.osm.xml', [114.03814, 22.51675, 114.06963, 22.56533])
-    download_osm_xml('../../cache/Shenzhen.osm.xml', SZ_BBOX)
     download_osm_xml('../../cache/GBA.osm.xml', GBA_BBOX)
     
     proj_name = "GaoxinParkMiddle"